Move the NES Go example to the package-directory layout
PR #155 made Go library output declare `package {module-name}` and
moved the DOOM example to the matching layout (generated file in a
subdirectory, frontend inside the package, a thin main.go), but the
NES example, merged one day earlier in #115, was left behind. Since
then its build.sh generated `package nes` beside `package main` and
`go build` failed with "found packages main and nes".

Mirror the DOOM layout: nes_gen.go is generated into nes/, the
frontend moves to nes/host.go (it reads the unexported
nesInst.memory.data, so it has to be a file in the package), main.go
shrinks to importing the package and calling nes.Run(), and the
binary moves to bin/nes because ./nes is now the package directory.
The default ROM path walks one more directory up to follow the
binary.
