+++ examples/nes/go/.gitignore
@@ -1,3 +1,3 @@
-/nes_gen.go
-/nes
+/nes/nes_gen.go
+/bin/
 /screenshot.png
