+++ examples/nes/go/build.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
-# Regenerate the dewasm-generated NES library and build the ebiten frontend.
-# nes_gen.go is generated code and is gitignored, so this step has to run before `go build` can succeed from a clean checkout.
+# Regenerate the dewasm-generated NES library and build the ebiten frontend. nes/nes_gen.go is generated code and is gitignored, so this step has to run before `go build` can succeed from a clean checkout.
+# The binary goes to bin/nes: the package directory beside it is already called nes.
 set -euo pipefail
 cd "$(dirname "$0")"
 
@@ -13,9 +13,9 @@ repo_root="$(cd ../../.. && pwd)"
   cargo run -q -p dewasm-cli -- \
     examples/apps/cache/nes.wasm \
     --target go --mode library --module-name nes \
-    -o examples/nes/go/nes_gen.go
+    -o examples/nes/go/nes/nes_gen.go
 )
 
-go build -o nes .
+go build -o bin/nes .
 
-echo "built $(pwd)/nes"
+echo "built $(pwd)/bin/nes"
